cleanup!: move FeedForward to model definition files
This `nn.Module` is opinionated and not flexible. It only
offers GEGLU-based FFN which was used in POYO and POYO+.
FFNs of all different varieties exist and different variants
are regularly used by ML researchers. Thus, it would be best
if we do not offer a FFN at all in our public interface.
It is standard for users to write their own FFNs.
