+++ torch_brain/models/poyo.py
@@ -5,13 +5,13 @@ from pathlib import Path
 import numpy as np
 import torch
 import torch.nn as nn
+import torch.nn.functional as F
 from torchtyping import TensorType
 from temporaldata import Data
 
 from torch_brain.data import chain, pad8, track_mask8
 from torch_brain.nn import (
     Embedding,
-    FeedForward,
     InfiniteVocabEmbedding,
     RotaryCrossAttention,
     RotarySelfAttention,
@@ -437,3 +437,35 @@ def poyo_mp(readout_spec: ModalitySpec, ckpt_path=None):
         t_min=1e-4,
         t_max=4.0,
     )
+
+
+class _GEGLU(nn.Module):
+    """Gated Gaussian Error Linear Unit (GEGLU) activation function, as introduced in
+    the paper "GLU Variants Improve Transformer" (https://arxiv.org/abs/2002.05202).
+    """
+
+    def forward(self, x):
+        x, gates = x.chunk(2, dim=-1)
+        return x * F.gelu(gates)
+
+
+class FeedForward(nn.Module):
+    """A feed-forward network with GEGLU activation.
+
+    Args:
+        dim (int): Input and output dimension
+        mult (int, optional): Multiplier for hidden dimension. Defaults to 4
+        dropout (float, optional): Dropout probability. Defaults to 0.2
+    """
+
+    def __init__(self, dim, mult=4, dropout=0.2):
+        super().__init__()
+        self.net = nn.Sequential(
+            nn.Linear(dim, dim * mult * 2),
+            _GEGLU(),
+            nn.Dropout(p=dropout),
+            nn.Linear(dim * mult, dim),
+        )
+
+    def forward(self, x):
+        return self.net(x)
