+++ torch_brain/models/poyo_plus.py
@@ -3,13 +3,13 @@ import logging
 
 import numpy as np
 import torch.nn as nn
+import torch.nn.functional as F
 from torchtyping import TensorType
 from temporaldata import Data
 
 from torch_brain.data import chain, pad8, track_mask8
 from torch_brain.nn import (
     Embedding,
-    FeedForward,
     InfiniteVocabEmbedding,
     MultitaskReadout,
     RotaryCrossAttention,
@@ -376,3 +376,35 @@ class POYOPlus(nn.Module):
                 f"sequence_length ({sequence_length}) is not a multiple of latent_step "
                 f"({latent_step}). This is a simple warning, and this behavior is allowed."
             )
+
+
+class _GEGLU(nn.Module):
+    """Gated Gaussian Error Linear Unit (GEGLU) activation function, as introduced in
+    the paper "GLU Variants Improve Transformer" (https://arxiv.org/abs/2002.05202).
+    """
+
+    def forward(self, x):
+        x, gates = x.chunk(2, dim=-1)
+        return x * F.gelu(gates)
+
+
+class FeedForward(nn.Module):
+    """A feed-forward network with GEGLU activation.
+
+    Args:
+        dim (int): Input and output dimension
+        mult (int, optional): Multiplier for hidden dimension. Defaults to 4
+        dropout (float, optional): Dropout probability. Defaults to 0.2
+    """
+
+    def __init__(self, dim, mult=4, dropout=0.2):
+        super().__init__()
+        self.net = nn.Sequential(
+            nn.Linear(dim, dim * mult * 2),
+            _GEGLU(),
+            nn.Dropout(p=dropout),
+            nn.Linear(dim * mult, dim),
+        )
+
+    def forward(self, x):
+        return self.net(x)
